Move searchTermsToAlternation to UpdateHeuristic
